% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twc-package.R
\docType{package}
\name{twc-package}
\alias{twc}
\alias{twc-package}
\title{twc: Terrestrial Water Cycle}
\description{
twc is an R package that that constitutes the core functions to analyze terrestrial water cycle data across various spatio-temporal scales.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/imarkonis/twc}
  \item Report bugs at \url{https://github.com/imarkonis/twc/issues}
}

}
\author{
\strong{Coded by:} Mijael Rodrigo Vargas Godoy \email{vargas_godoy@fzp.czu.cz}

\strong{Conceptual design by:} Mijael Rodrigo Vargas Godoy \email{vargas_godoy@fzp.czu.cz} and Yannis Markonis \email{markonis@fzp.czu.cz}
}
\keyword{internal}
